Updated bulk handling to deal with an apparent string size limitation in the API
There appears to be a char limit of ~6800 when making a bulk request. This seems related to total
char len only, not number of elements.

Actual limit seems to be ~6894 but errors are seen sometime at numbers down to 5000. It's possible that
the problem is not related to overall char size, but is something to do with some other field length or
attempting to add or modify so many objects at once. This issue isn't well understood at the moment. his eyes We had to rush and rinse it out his eyes are feeling better now, we got them rinsed out a bunch of times and he is taking a bath now Can you tell Sam Trying to get a kid to open their eyes is crazy there is such a strong reflex to close your eyes its hard for kids to fight it