Updated the put method to wrap all input params in update.
for example:

{'updates': {'status':'inp'}} vs {'status':'inp'}

This was needed because the API cannot handle nested lists as dict param values. This occurs when trying to update a collection. For example, updating the custom forms attached to a task. While this fix isn't needed for every call it was easier than having to look through each params to see if it contained a list.